+++ src/main/kotlin/com/beanz/drpings/DRPingsPanel.kt
@@ -1,0 +1,23 @@
+package com.beanz.drpings
+
+import com.mumfrey.liteloader.core.LiteLoader
+import com.mumfrey.liteloader.modconfig.AbstractConfigPanel
+import com.mumfrey.liteloader.modconfig.ConfigPanelHost
+import net.minecraft.client.gui.GuiButton
+
+class DRPingsPanel : AbstractConfigPanel() {
+  override fun getPanelTitle() = "DR Pings Config"
+
+
+
+  override fun onPanelHidden() {
+
+    LiteLoader.getInstance().writeConfig(LiteModDRPings.mod)
+  }
+
+  override fun addOptions(host: ConfigPanelHost) {
+
+
+
+  }
+}
