--- src/main/resources/mixin.drimprovement.json
@@ -1,18 +1,0 @@
-{
-  "required": true,
-  "minVersion": "0.4.10",
-  "package": "com.beanz.drpings.mixins",
-  "refmap": "mixin.drpings.refmap.json",
-  "compatibilityLevel": "JAVA_8",
-  "mixins": [
-    "AbstractClientPlayerMixin",
-    "EntityPlayerSpMixin",
-    "GuiBossOverlayMixin",
-    "GuiIngameMixin",
-    "GuiNewChatMixin",
-    "ItemRendererMixin",
-    "ItemStackMixin",
-    "MinecraftMixin",
-    "NetworkManagerMixin"
-  ]
-}
